Criacao do projeto spring boot com maven, configuracao de persistencia com JPA, Configuracao de acesso ao banco de dados, adicao de validacao de dados e  preparacao para a aplicacao para WEB / Api

+++ sigfebackend/src/main/java/com/sigfe/backend/model/Compra.java
@@ -14,7 +14,7 @@ public class Compra extends Transacao {
     private FormaPagamento formaPagamento;
     private StatusCompra status;
     private String numeroDocumento;
-    public Compra(int id, BigDecimal valorTotal, List<ItemTransacao> itens, String usuario,
+    public Compra(Long id, BigDecimal valorTotal, List<ItemTransacao> itens, String usuario,
                   Fornecedor fornecedor, FormaPagamento formaPagamento, String numeroDocumento) {
         super(id, valorTotal, itens, usuario);
         this.fornecedor = fornecedor;
