+++ sigfebackend/src/main/java/com/sigfe/backend/model/Venda.java
@@ -5,7 +5,7 @@ import java.math.BigDecimal;
 import java.util.List;
 
 public class Venda extends Transacao {
-    public Venda(int id, BigDecimal valorTotal, List<ItemTransacao> itens, String usuario)
+    public Venda(Long id, BigDecimal valorTotal, List<ItemTransacao> itens, String usuario)
     {
         super(id, valorTotal, itens, usuario);
     }
